create ecs folder and move components to it
