+++ src/lib/canvas.js
@@ -1,6 +1,6 @@
 import _ from "lodash";
 import * as PIXI from "pixi.js";
-import { Position } from "../components";
+import { Position } from "../ecs/components";
 import { getEntityData } from "./ecsHelpers";
 import { grid } from "./grid";
 import * as gfx from "./graphics";
