+++ src/prefabs/sword.js
@@ -11,7 +11,7 @@ import {
   Position,
   Wieldable,
   Zindex,
-} from "../components";
+} from "../ecs/components";
 
 import { addSprite } from "../lib/canvas";
 import { updatePosition } from "../lib/ecsHelpers";
