+++ src/ui/legend.js
@@ -2,7 +2,7 @@ import _ from "lodash";
 import { clearContainer, printRow, printTemplate } from "../lib/canvas";
 import { distance } from "../lib/grid";
 import { getEntityData } from "../lib/ecsHelpers";
-import { Health, Position } from "../components";
+import { Health, Position } from "../ecs/components";
 
 const container = "legend";
